Make the Version class throw a checked exception
The Version class was implemented as a representation of the app's
versioning scheme. However, it's constructor was declared to throw the
unchecked NumberFormatException in the case where it encountered a non-
numerical character while parsing the tokens. Since this is a case that
does need to be handled at runtime, this commit changes the
implementation to throw a checked ParseException instead. The client
code at NewVersionBroadcastReceiver is implemented to abort operation
upon encountering this exception, since this would be an unexpected and
inconsistent state.