Document token replacement in app store config
The 'application_id' token in the app store URIs config list is replaced
at runtime by the actual application ID. @miankhalid suggested that this
should be documented. Since there isn't really much documentation of the
config properties in this project, and since the application ID token
replacement is done internally, this commit just adds an inner comment
on the operation (though it's mostly self-evident anyway).