Simplify invalid new version event data handling
Previously, the arguments were validated in the constructor of the
NewVersionAvailableEvent event class, which threw an
IllegalStateException in the case where all the arguments were null or
otherwise empty of information. This exception was then caught (as an
IllegalArgumentException, which was what was documented and should have
been thrown instead) and suppressed in the static post() method, as per
it's documentation. Then the NewVersionBroadcastInterceptor class added
it's own validation checks before calling the post() method, which
included a comparison against the current build's version number as
well. Thus, there was lots of redudancy and confusion in the logic.

This commit aims to reform this situation, by centralizing all the
validation (including the version number comparison) in the event
constructor with appropriate inline comments, extending the javadoc of
both the constructor and the static post() method to include the new
comparison validity check, and adding an inline comment in the post()
method where it suppresses the constructor's thrown exceptions.